bot_core/requirements.txt
# Pulsareon Sensory Dependencies
edge-tts
openai-whisper
sounddevice
numpy
scipy
chardet
requests
psutil
sentence-transformers
torch
opencv-python
